+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Added offline for product batch list and detail
+- Added a Capacitations view for Disabilities List
 
 ## [0.2.0] - 29/11/2025